refactor: update account confirmation page
